+++ e2e/tests-dfx/install.bash
@@ -108,6 +108,20 @@ teardown() {
   assert_match 'hello-script'
 }
 
+@test "post-install tasks run in project root" {
+  install_asset post_install
+  dfx_start
+
+  assert_command dfx canister create --all
+  assert_command dfx build
+
+  cd src/e2e_project_backend
+
+  assert_command dfx canister install postinstall_script
+  assert_match 'hello-script'
+  assert_match "working directory of post-install script: '.*/working-dir/e2e_project'"
+}
+
 @test "post-install tasks receive environment variables" {
   install_asset post_install
   dfx_start
@@ -324,4 +338,29 @@ Please remove one of them or leave both undefined."
   do
   echo yes | dfx canister install --mode=reinstall custom
   done
+}
+
+@test "specify upgrade options (skip_pre_upgrade, wasm_memory_persistence)" {
+  dfx_start
+  dfx canister create e2e_project_backend
+  dfx build e2e_project_backend
+
+  # The canister is not installed yet and the mode is 'auto'.
+  # The actual InstallMode will be 'install'.
+  # In this case, the provided upgrade options are just hint which doesn't take effect.
+  assert_command dfx canister install e2e_project_backend --mode auto --skip-pre-upgrade
+
+  assert_command dfx canister install e2e_project_backend --mode auto --wasm-memory-persistence keep
+  assert_command dfx canister install e2e_project_backend --mode auto --wasm-memory-persistence replace
+  assert_command dfx canister install e2e_project_backend --mode auto --skip-pre-upgrade --wasm-memory-persistence keep
+
+  assert_command dfx canister install e2e_project_backend --mode upgrade --skip-pre-upgrade
+  assert_command dfx canister install e2e_project_backend --mode upgrade --wasm-memory-persistence keep
+  assert_command dfx canister install e2e_project_backend --mode upgrade --wasm-memory-persistence replace
+  assert_command dfx canister install e2e_project_backend --mode upgrade --skip-pre-upgrade --wasm-memory-persistence keep
+
+  assert_command_fail dfx canister install e2e_project_backend --mode install --skip-pre-upgrade
+  assert_contains "--skip-pre-upgrade and --wasm-memory-persistence can only be used with mode 'upgrade' or 'auto'."
+  assert_command_fail dfx canister install e2e_project_backend --mode reinstall --wasm-memory-persistence keep
+  assert_contains "--skip-pre-upgrade and --wasm-memory-persistence can only be used with mode 'upgrade' or 'auto'."
 }