+++ scripts/workflows/provision-linux.sh
@@ -51,8 +51,8 @@ if [ "$E2E_TEST" = "tests-dfx/build_rust.bash" ]; then
 fi
 
 if [ "$E2E_TEST" = "tests-dfx/deps.bash" ]; then
-     cargo install cargo-binstall@1.6.9
-     cargo binstall -y ic-wasm
+     cargo install cargo-binstall@1.6.9 --locked
+     cargo binstall -y ic-wasm --locked
 fi
 
 if [ "$E2E_TEST" = "tests-icx-asset/icx-asset.bash" ]; then