+++ src/dfx-core/src/extension/manager/install.rs
@@ -4,6 +4,7 @@ use crate::error::extension::{
     GetHighestCompatibleVersionError, GetTopLevelDirectoryError, InstallExtensionError,
 };
 use crate::extension::{
+    catalog::ExtensionCatalog,
     manager::ExtensionManager,
     manifest::{ExtensionDependencies, ExtensionManifest},
     url::ExtensionJsonUrl,
@@ -29,11 +30,21 @@ pub enum InstallOutcome {
 impl ExtensionManager {
     pub async fn install_extension(
         &self,
-        url: &ExtensionJsonUrl,
+        name: &str,
+        catalog_url: Option<&Url>,
         install_as: Option<&str>,
         version: Option<&Version>,
     ) -> Result<InstallOutcome, InstallExtensionError> {
-        let manifest = Self::get_extension_manifest(url).await?;
+        let url = if let Ok(url) = Url::parse(name) {
+            ExtensionJsonUrl::new(url)
+        } else {
+            ExtensionCatalog::fetch(catalog_url)
+                .await?
+                .lookup(name)
+                .ok_or(InstallExtensionError::ExtensionNotFound(name.to_string()))?
+        };
+
+        let manifest = Self::get_extension_manifest(&url).await?;
         let extension_name: &str = &manifest.name;
 
         let effective_extension_name = install_as.unwrap_or(extension_name);
@@ -59,7 +70,7 @@ impl ExtensionManager {
 
         let extension_version = match version {
             Some(version) => version.clone(),
-            None => self.get_highest_compatible_version(url).await?,
+            None => self.get_highest_compatible_version(&url).await?,
         };
         let github_release_tag = get_git_release_tag(extension_name, &extension_version);
         let extension_archive = get_extension_archive_name(extension_name)?;
@@ -153,6 +164,7 @@ impl ExtensionManager {
         Ok(temp_dir)
     }
 
+    #[allow(clippy::collapsible_if)]
     fn finalize_installation(
         &self,
         extension_name: &str,
@@ -162,17 +174,23 @@ impl ExtensionManager {
         let effective_extension_dir = &self.get_extension_directory(effective_extension_name);
         let top_level_dir = get_top_level_directory(temp_dir.path())?;
         crate::fs::rename(&top_level_dir, effective_extension_dir)?;
-        if extension_name != effective_extension_name {
-            // rename the binary
-            crate::fs::rename(
-                &effective_extension_dir.join(extension_name),
-                &effective_extension_dir.join(effective_extension_name),
-            )?;
-        }
-        #[cfg(unix)]
+
+        let installed_manifest = ExtensionManifest::load(effective_extension_name, &self.dir)?;
+
+        if matches!(installed_manifest.subcommands, Some(subcommands) if !subcommands.0.is_empty())
         {
-            let bin = effective_extension_dir.join(effective_extension_name);
-            crate::fs::set_permissions(&bin, std::fs::Permissions::from_mode(0o500))?;
+            if extension_name != effective_extension_name {
+                // rename the binary
+                crate::fs::rename(
+                    &effective_extension_dir.join(extension_name),
+                    &effective_extension_dir.join(effective_extension_name),
+                )?;
+            }
+            #[cfg(unix)]
+            {
+                let bin = effective_extension_dir.join(effective_extension_name);
+                crate::fs::set_permissions(&bin, std::fs::Permissions::from_mode(0o500))?;
+            }
         }
         Ok(())
     }