+++ src/dfx/src/lib/operations/canister/install_canister.rs
@@ -9,6 +9,7 @@ use crate::lib::operations::canister::all_project_canisters_with_ids;
 use crate::lib::operations::canister::motoko_playground::authorize_asset_uploader;
 use crate::lib::state_tree::canister_info::read_state_tree_canister_module_hash;
 use crate::util::assets::wallet_wasm;
+use crate::util::clap::install_mode::InstallModeHint;
 use crate::util::{blob_from_arguments, get_candid_init_type, read_module_metadata};
 use anyhow::{anyhow, bail, Context};
 use backoff::backoff::Backoff;
@@ -21,7 +22,7 @@ use dfx_core::config::model::network_descriptor::NetworkDescriptor;
 use dfx_core::identity::CallSender;
 use fn_error_context::context;
 use ic_agent::Agent;
-use ic_utils::interfaces::management_canister::builders::InstallMode;
+use ic_utils::interfaces::management_canister::builders::{InstallMode, WasmMemoryPersistence};
 use ic_utils::interfaces::ManagementCanister;
 use ic_utils::Argument;
 use itertools::Itertools;
@@ -42,7 +43,7 @@ pub async fn install_canister(
     wasm_path_override: Option<&Path>,
     argument_from_cli: Option<&str>,
     argument_type_from_cli: Option<&str>,
-    mode: Option<InstallMode>,
+    mode_hint: &InstallModeHint,
     call_sender: &CallSender,
     upgrade_unchanged: bool,
     pool: Option<&CanisterPool>,
@@ -63,13 +64,37 @@ pub async fn install_canister(
         "Previously installed module hash: {:?}",
         installed_module_hash.as_ref().map(hex::encode)
     );
-    let mode = mode.unwrap_or_else(|| {
-        if installed_module_hash.is_some() {
-            InstallMode::Upgrade(None)
-        } else {
-            InstallMode::Install
-        }
-    });
+    let wasm_memory_persistence_embeded =
+        read_module_metadata(agent, canister_id, "enhanced-orthogonal-persistence")
+            .await
+            .map(|_| WasmMemoryPersistence::Keep);
+    // let mode = if let InstallModeHint::Auto(options) = mode_hint {
+    //     if installed_module_hash.is_some() {
+    //         InstallMode::Upgrade(Some(CanisterUpgradeOptions {
+    //             wasm_memory_persistence,
+    //             skip_pre_upgrade: None,
+    //         }))
+    //     } else {
+    //         InstallMode::Install
+    //     }
+    // } else {
+    //     InstallMode::Install
+    // };
+    let mode = mode_hint.to_install_mode(
+        installed_module_hash.is_some(),
+        wasm_memory_persistence_embeded,
+    );
+
+    // let mode = mode.unwrap_or_else(|| {
+    //     if installed_module_hash.is_some() {
+    //         InstallMode::Upgrade(Some(CanisterUpgradeOptions {
+    //             wasm_memory_persistence,
+    //             skip_pre_upgrade: None,
+    //         }))
+    //     } else {
+    //         InstallMode::Install
+    //     }
+    // });
     let mode_str = install_mode_to_prompt(&mode);
     let canister_name = canister_info.get_name();
     info!(
@@ -96,9 +121,13 @@ pub async fn install_canister(
         let stable_types = read_module_metadata(agent, canister_id, "motoko:stable-types").await;
         if let Some(stable_types) = &stable_types {
             match check_stable_compatibility(canister_info, env, stable_types) {
-                Ok(None) => (),
-                Ok(Some(err)) => {
-                    let msg = format!("Stable interface compatibility check failed for canister '{}'.\nUpgrade will either FAIL or LOSE some stable variable data.\n\n", canister_info.get_name()) + &err;
+                Ok(StableCompatibility::Okay) => (),
+                Ok(StableCompatibility::Warning(details)) => {
+                    let msg = format!("Stable interface compatibility check issued a WARNING for canister '{}'.\n\n", canister_info.get_name()) + &details;
+                    ask_for_consent(&msg)?;
+                }
+                Ok(StableCompatibility::Error(details)) => {
+                    let msg = format!("Stable interface compatibility check issued an ERROR for canister '{}'.\nUpgrade will either FAIL or LOSE some stable variable data.\n\n", canister_info.get_name()) + &details;
                     ask_for_consent(&msg)?;
                 }
                 Err(e) => {
@@ -366,11 +395,17 @@ async fn wait_for_module_hash(
     Ok(())
 }
 
+enum StableCompatibility {
+    Okay,
+    Warning(String),
+    Error(String),
+}
+
 fn check_stable_compatibility(
     canister_info: &CanisterInfo,
     env: &dyn Environment,
     stable_types: &str,
-) -> anyhow::Result<Option<String>> {
+) -> anyhow::Result<StableCompatibility> {
     use crate::lib::canister_info::motoko::MotokoCanisterInfo;
     let info = canister_info.as_info::<MotokoCanisterInfo>()?;
     let stable_path = info.get_output_stable_path();
@@ -387,12 +422,16 @@ fn check_stable_compatibility(
         .arg("--stable-compatible")
         .arg(&deployed_stable_path)
         .arg(stable_path)
+        .current_dir(canister_info.get_workspace_root())
         .output()
         .context("Failed to run 'moc'.")?;
+    let message = String::from_utf8_lossy(&output.stderr).to_string();
     Ok(if !output.status.success() {
-        Some(String::from_utf8_lossy(&output.stderr).to_string())
+        StableCompatibility::Error(message)
+    } else if !output.stderr.is_empty() {
+        StableCompatibility::Warning(message)
     } else {
-        None
+        StableCompatibility::Okay
     })
 }